Fix Update Button Bug
The update button only worked for newly created replacement items and would throw an error when it was used with a part from the sample inventory. I updated the BootStrapData class to create new Outsourced part objects to store those objects as the sample inventory in the outsourced part repo. This fixed the bug and the code now works as intended.